pyproject.toml
[tool.poetry]
name = "langgraph-test"
version = "0.1.0"
description = ""
authors = ["Jayasimha Raghavan <mr.jayasimha@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.25"
python-dotenv = "^1.1.0"
langchain-openai = "^0.3.12"
langsmith = "^0.3.42"
langchain-core = "^0.3.58"
langgraph = "^0.4.3"
matplotlib = "^3.8.2"
networkx = "^3.2.1"
numpy = "^1.26.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
